add a base class for smooth descriptors
